+++ R/auto_prior.R
@@ -171,16 +171,28 @@ prepare_prior <- function(prior, data, stan_data, model, pooling, covariates,
               message("Setting hyper-SD prior using 10 times the naive SD across sites")
               special_name <- "sigma_tau"
             }
-          } else if(current_prior == "control" && model == "logit") {
-            prop_ctrl <- data$c / (data$c + data$d)
-            if(max(prop_ctrl) > .999 | min(prop_ctrl) < .001)
-              message("Baseline proportion of events is very close to 0 or 1.",
-                      "Consider manually setting prior_control.")
-            special_name <- "log odds of event rate in untreated: mean"
-          } else if(current_prior == "control_sd" && model == "logit") {
-            if(stan_data$pooling_baseline != 0)
-              special_name <- "log odds of event rate in untreated: sd"
-            else
+          } else if(current_prior == "control") {
+            if(model == "logit"){
+              prop_ctrl <- data$c / (data$c + data$d)
+              if(max(prop_ctrl) > .999 | min(prop_ctrl) < .001)
+                message("Baseline proportion of events is very close to 0 or 1.",
+                        "Consider manually setting prior_control.")
+              special_name <- "log odds of event rate in untreated: mean"
+            }
+            if(model %in% c("rubin_full", "mutau_full")){
+              if(stan_data$pooling_baseline != 0)
+                special_name <- "mu (hyperparameter)"
+              else
+                special_name <- "independent prior on control means"
+            }
+          } else if(current_prior == "control_sd") {
+            if(stan_data$pooling_baseline != 0){
+              if(model == "logit")
+                special_name <- "log odds of event rate in untreated: sd"
+              if(model %in% c("rubin_full", "mutau_full")){
+                special_name <- "sigma_mu (hyperparameter)"
+              }
+            }else
               special_name <- "DNP" #do not print, this is not used!
           }
 
@@ -198,7 +210,7 @@ prepare_prior <- function(prior, data, stan_data, model, pooling, covariates,
              pooling != "partial")
             message("Prior for hyper-SD set, but pooling is not partial. Ignoring.")
           if(current_prior == "control_sd" &&
-             model == "logit" &&
+             model %in% c("logit", "rubin_full", "mutau_full") &&
              stan_data$pooling_baseline == 0)
             message("SD hyperparameter for control groups defined,",
                     "but there is no pooling. Ignoring it.")
@@ -211,11 +223,11 @@ prepare_prior <- function(prior, data, stan_data, model, pooling, covariates,
   }
 
   if(model == "quantiles") {
-    prior_list[["prior_dispersion_on_beta_0"]] <- 1000*diag(stan_data$Nq)
-    prior_list[["prior_dispersion_on_beta_1"]] <- 1000*diag(stan_data$Nq)
-    if(!silent) {
-      message("Sigma (hypervariance-covariance) = 1000*I_{Nquantiles}")
-    }
+    # prior_list[["prior_dispersion_on_beta_0"]] <- 1000*diag(stan_data$Nq)
+    # prior_list[["prior_dispersion_on_beta_1"]] <- 1000*diag(stan_data$Nq)
+    # if(!silent) {
+    #   message("Sigma (hypervariance-covariance) = 1000*I_{Nquantiles}")
+    # }
   }
 
   # Setting covariates prior